// webApp/backend/src/services/authService/types.ts
//==========================================================================================================
//? Types
//==========================================================================================================

type AuthServiceResult<T = undefined> = {
    status: number;
    messageKey: string;
    data?: T;
};

export type { AuthServiceResult };